Rename __common.h to sassl_common.h
