+++ external/ossl_lib/include/asm/unistd.h
@@ -1,7 +1,7 @@
 #ifndef _UNISTD_H
 #define _UNISTD_H
 
-#include "__common.h"
+#include "sassl_common.h"
 
 static inline int close(int fd)
 {
