+++ external/ossl_lib/include/limits.h
@@ -1,4 +1,4 @@
-#include "__common.h"
+#include "sassl_common.h"
 #define INT_MAX 0x7fffffff
 #define UINT_MAX 0xffffffffU
 #define LONG_MAX 0x7fffffff
