+++ external/ossl_lib/include/stdarg.h
@@ -1,4 +1,4 @@
-#include "__common.h"
+#include "sassl_common.h"
 #define va_list __builtin_va_list
 #define va_start __builtin_va_start
 #define va_arg __builtin_va_arg
