+++ external/ossl_lib/include/stdatomic.h
@@ -1,1 +1,1 @@
-#include "__common.h"
+#include "sassl_common.h"
