+++ external/ossl_lib/include/stddef.h
@@ -1,4 +1,4 @@
-#include "__common.h"
+#include "sassl_common.h"
 
 #define offsetof(type, member) __builtin_offsetof(type, member)
