+++ external/ossl_lib/include/stdio.h
@@ -1,7 +1,7 @@
 #ifndef _STDIO_H
 #define _STDIO_H
 
-#include "__common.h"
+#include "sassl_common.h"
 #include <stdarg.h>
 
 #define BUFSIZ 1024
