+++ external/ossl_lib/include/sys/param.h
@@ -1,1 +1,1 @@
-#include "../__common.h"
+#include "../sassl_common.h"
