+++ external/ossl_lib/include/sys/stat.h
@@ -1,7 +1,7 @@
 #ifndef _STAT_H
 #define _STAT_H
 
-#include "../__common.h"
+#include "../sassl_common.h"
 #include <time.h>
