+++ external/ossl_lib/include/sys/types.h
@@ -1,2 +1,2 @@
-#include "../__common.h"
+#include "../sassl_common.h"
