+++ external/ossl_lib/include/time.h
@@ -1,7 +1,7 @@
 #ifndef _TIME_H
 #define _TIME_H
 
-#include "__common.h"
+#include "sassl_common.h"
 
 struct tm
 {
